requirements.txt
aws-cdk-lib==2.15.0
constructs>=10.0.0,<11.0.0
discord.py==1.7.3
gitlab==1.0.2
python-dotenv==0.19.2
python-gitlab==3.2.0
